+++ CHANGELOG.md
@@ -28,7 +28,7 @@
   - `discover-hostname` - discover hostname and populate config
   - `discover-project-edn-version` - discover project version and populate config
   - `fallback-config-files` - specify config filenames when unspecified
-  - `update-launcher` - specify alternate launcher on-the-fly
+  - `prepare-launcher` - specify alternate launcher on-the-fly
 - Fix issue where resolving new config ignores existing config